python-backend/requirements.txt
numpy 
Flask
Keras 
tensorflow
pillow